changed to use graph class added better colors, added panning

+++ app/src/main/java/graph/Vertex.java
@@ -2,9 +2,9 @@ package graph;
 
 public class Vertex {
     int id;
-    int x;
-    int y;
-    public Vertex(int id, int x, int y) {
+    double x;
+    double y;
+    public Vertex(int id, double x, double y) {
         this.id = id;
         this.x = x;
         this.y = y;
